Move async upload test to separate class
