+++ src/test/java/com/czertainly/core/messaging/jms/listeners/JmsListenerIntegrationTest.java
@@ -1,28 +1,21 @@
 package com.czertainly.core.messaging.jms.listeners;
 
-import com.czertainly.api.model.client.certificate.UploadCertificateRequestDto;
 import com.czertainly.api.model.core.auth.Resource;
 import com.czertainly.api.model.core.other.ResourceEvent;
-import com.czertainly.core.dao.repository.CertificateRepository;
-import com.czertainly.core.helpers.CertificateGeneratorHelper;
 import com.czertainly.core.messaging.jms.producers.EventProducer;
 import com.czertainly.core.messaging.model.EventMessage;
-import com.czertainly.core.service.CertificateService;
 import com.czertainly.core.util.BaseMessagingIntTest;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 
-import java.security.cert.X509Certificate;
-import java.util.Base64;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doAnswer;
 import static org.mockito.Mockito.verify;
@@ -48,17 +41,11 @@ class JmsListenerIntegrationTest extends BaseMessagingIntTest {
     @Autowired
     private EventProducer eventProducer;
 
-    @Autowired
-    private CertificateService certificateService;
-
-    @Autowired
-    private CertificateRepository certificateRepository;
-
     // @MockitoBean (not @MockitoSpyBean) is intentional: EventListener is @Transactional, so Spring
     // wraps it in an AOP proxy. A spy wraps the underlying bean but the AOP proxy (injected into
     // EventJmsEndpointConfig.listenerMessageProcessor) bypasses the spy. @MockitoBean replaces the
     // bean entirely — no AOP proxy — so doAnswer intercepts the call directly.
-    @Autowired
+    @MockitoBean
     private EventListener eventListener;
 
     @Test
@@ -88,17 +75,4 @@ class JmsListenerIntegrationTest extends BaseMessagingIntTest {
         verify(eventListener).processMessage(any(EventMessage.class));
     }
 
-    @Test
-    void testCertificateUploadedEvent() throws Exception {
-        X509Certificate certificate = CertificateGeneratorHelper.generateCACertificate(null, "CN=TestCA");
-        String content = Base64.getEncoder().encodeToString(certificate.getEncoded());
-        UploadCertificateRequestDto request = new UploadCertificateRequestDto();
-        request.setCertificate(content);
-        String fingerprint = certificateService.uploadAsync(request).getFingerprint();
-        assertThat(fingerprint).isNotNull();
-        await().atMost(10, TimeUnit.SECONDS).until(() ->
-                certificateRepository.findByFingerprint(fingerprint).isPresent()
-        );
-        Assertions.assertTrue(certificateRepository.findByFingerprint(fingerprint).isPresent());
-    }
 }
